+++ pokedex/src/components/PokemonCard/styled.js
